sched: ucassist: Reduce soft IRQ spam by using a timestamp
Instead of having the input trigger queue a soft IRQ everytime, let's
have it configure the input timestamp instead whenever the input timer
is running. This is to avoid flooding the CPU with IRQs which can cause
jitter to increase.

Using this timestamp, we can evaluate at timer expiration whether the
timestamp is newer or the same as the timer's timestamp, and trigger
an update without using a soft IRQ since we're able to perform kthread
queues within the timer function.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>