sched: ucassist: Properly lock the task RQ when checking user_defined
Prevent multiple threads from overwriting a target's values unnecessarily
or data racing by putting the user_defined check under the task rq lock.
This is okay as the check shouldn't take long, and that this slight
overhead is worth over wasting CPU time rewriting the same values when
we've already done it once, as well as avoiding possible data races with
other threads when we read it unprotected.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>