sched: ucassist: Move active input timer updates completely in the timer function
We don't need to queue the kthread work just to renew the timer, when
we can perform this action at the timer function itself without causing
significant overhead. Also remove timer_pending as we can just simply
use the sleep state to protect the IRQ queue. With this, restrict the
kthread from renewing the timer to avoid causing any contention between
the timer and the kthread, and to serialize the update without using
locks.

Also fix timestamp not being updated at the fb notifier.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>