sched: ucassist: Reduce input event timeout to 100ms
This is 20ms longer than the 80ms idle timer by the GPU. We don't
need to keep the CPU at a higher load once the renderthread has
stopped writing to the GPU IOCTL for 100ms after the last write as
the GPU has likely gone idle.

Also remove unused sleep_states log at the FB notifier and switch
the sleep_state log in the update_fn to pr_debug.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>