arm64: DT: msm8998: Configure all known IRQ types according to mainline
This is to make sure HLOS know the configuration of at least most of
the IRQs in case the bootloader or their drivers did not configure
them. This is in-line with mainline convention wherein IRQ types
must be defined for the IRQ controller to prevent any unknown
configuration and behavior.

Also remove any intc parents other than BLSP and MPM irqs to configure
MPM first before passing over to GICv3.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>