arm64: DT: msm8998: Add AGGR1 and AGGR2 NOC voter clocks and update configuration
sdm660 has introduced voter clocks for AGGR2 NOC clock as to handle multiple
client votes. This is also the same case with msm8998 as we vote on these
AGGR NOC clocks on various drivers, particularly the MSM BUS driver and
the SMMU driver can vote on these clocks at the same time along with various
drivers. This will ensure that the voted rate is aggregated and proper
according to which client is voting, and that its state is properly
handled by the clock framework.

Furthermore, their bus configuration has been updated to remove msm8998
configurations applied at msmfalcon as it was deemed unnecessary in
newer tags. Also remove unnecessary clock additions to GPU PIL TZ.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>