clk: qcom: clk-alpha-pll: Add support for Fabia PLLs to sync the CAL_L_VAL
This patch adds a flag which will have the clock driver sync the
PLL_L_VAL with the PLL_CAL_L_VAL to ensure stability for legacy
platforms such as msm8998. This also disables the calibration sequence
at prepare and the default calibration l value as we let the PLL handle
this, following downstream and ensuring we do not use the VCO rates
which are explicitly restricted for Fabia PLLs in the MSM clock driver.

Reference:
https://github.com/LineageOS/android_kernel_oneplus_msm8998/commit/85e3adb4575da5ddcd8b04b9d671ca5062b4528c

Co-authored-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>