clk: qcom: clk-branch: Allow configuration to restrict setting specific clock flags
In msm8998, we want to force the GFX peripheral to be always on to prevent
perf counter corruption, for this we set this CBCR val after clock
probe. However, after tracing the KGSL driver, it overwrites and toggles
this clock flag at clk_set_options whenever it enters SLUMBER state and
WAKE state - and there is no flag or code to prevent this from happening.

So, to prevent this from happening, as well as other possible occurences
in other drivers, let's introduce a new optional retain_cbcr value
for branch clocks to configure and restrict any writes towards specified
CBCR registers. This will block any driver from writing towards the
specified registers through the clock API.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>