clk: qcom: gpucc-msm8998: Complete branch clocks configuration
Let's complete all the required branch clock configurations for GPU
to ensure proper configuration.

Particularly, gpucc_cxo_clk (gpucc_xo) has been renamed and configured
completely as it was missing its enable_reg and enable_mask which can
cause issues when the hardware expects a vote on the branch. Also follow
newer QCOM SoCs configuration with the clk framework wherein we decouple the
gpucc_cxo_clk from the other clocks as we only need to vote on this
clock, use bi_tcxo_ao as a parent to the GPU clocks which require clock
rate scaling.

Another important addition is the gpucc_gfx3d_clk CBCR flag wherein we
want to force periph on always to avoid perf corruption, for this we
configure retain_cbcr to prevent the KGSL driver (and Linux entirely)
from toggling the register through the clock framework API and force
it at init by manually writing on the regmap.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>